% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppLaunch.R
\docType{data}
\name{AppLaunch}
\alias{AppLaunch}
\title{AppLaunch}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AppLaunch
}
\value{
methods to manipulate your app
}
\description{
A tools to launch app
}
\section{Methods}{

\describe{
  \item{\code{show_terminal}}{Display the terminal where your application is launched}
  \item{\code{stop}}{Stop your app}
  \item{\code{get_buffer}}{Get what you have in your terminal}
  \item{\code{get_url}}{Get the url of your application}
  \item{\code{open_app}}{Open the url of your application}
  \item{\code{restart}}{Restart your application}
  \item{\code{is_running}}{Test your terminal is still running}
  \item{\code{auto_restart}}{Monitors your directory and restarts your application if there is a change}
 }
}

\keyword{datasets}
